+++ src/Metadata/MetadataFactory.php
@@ -23,6 +23,8 @@ use Cline\Money\RationalMoney;
 use Cline\Numerus\Numerus;
 use Cline\PhoneNumber\PhoneNumber;
 use Cline\PostalCode\PostalCode;
+use Cline\SemVer\Constraint;
+use Cline\SemVer\Version;
 use Cline\Struct\AbstractData;
 use Cline\Struct\Attributes\AllowSuperfluousKeys;
 use Cline\Struct\Attributes\AllowUndefinedValues;
@@ -60,6 +62,7 @@ use Cline\Struct\Casts\MoneyCast;
 use Cline\Struct\Casts\NumerusCast;
 use Cline\Struct\Casts\PhoneNumberCast;
 use Cline\Struct\Casts\PostalCodeCast;
+use Cline\Struct\Casts\SemVerCast;
 use Cline\Struct\Contracts\CastInterface;
 use Cline\Struct\Contracts\ProvidesCastClassInterface;
 use Cline\Struct\Contracts\ProvidesItemValidationRulesInterface;
@@ -488,6 +491,10 @@ final class MetadataFactory
 
             if ($type === PostalCode::class) {
                 return PostalCodeCast::class;
+            }
+
+            if ($type === Version::class || $type === Constraint::class) {
+                return SemVerCast::class;
             }
 
             if (